+++ ca/django_ca/tests/test_management_actions.py
@@ -128,6 +128,28 @@ class ExtendedKeyUsageActionTestCase(ParserTestCaseMixin, TestCase):
             namespace.extended_key_usage,
         )
 
+    def test_deprecated_values(self) -> None:
+        """Test old, coma-separated list format."""
+
+        msg = r"^Passing a coma-separated list is deprecated, pass space-separated values instead\."
+        with self.assertRemovedIn126Warning(msg):
+            namespace = self.parser.parse_args(["--eku", "clientAuth,serverAuth"])
+        self.assertEqual(
+            x509.ExtendedKeyUsage([ExtendedKeyUsageOID.CLIENT_AUTH, ExtendedKeyUsageOID.SERVER_AUTH]),
+            namespace.extended_key_usage,
+        )
+
+    def test_deprecated_values_with_critical(self) -> None:
+        """Test old, coma-separated list format."""
+
+        msg = r"^Passing a coma-separated list is deprecated, pass space-separated values instead\."
+        with self.assertRemovedIn126Warning(msg):
+            namespace = self.parser.parse_args(["--eku", "critical,clientAuth,serverAuth"])
+        self.assertEqual(
+            x509.ExtendedKeyUsage([ExtendedKeyUsageOID.CLIENT_AUTH, ExtendedKeyUsageOID.SERVER_AUTH]),
+            namespace.extended_key_usage,
+        )
+
     def test_dotted_string_value(self) -> None:
         """Test passing a dotted string."""
         namespace = self.parser.parse_args(["--eku", "1.3.6.1.5.5.7.3.2"])
@@ -201,40 +223,44 @@ class KeyUsageActionTestCase(ParserTestCaseMixin, TestCase):
         namespace = self.parser.parse_args(["-k", "keyCertSign"])
         self.assertEqual(self.key_usage(key_cert_sign=True, critical=False).value, namespace.key_usage)
 
-        namespace = self.parser.parse_args(["-k", "keyCertSign,keyAgreement"])
+        namespace = self.parser.parse_args(["-k", "keyCertSign", "keyAgreement"])
         self.assertEqual(
             self.key_usage(key_cert_sign=True, key_agreement=True, critical=False).value, namespace.key_usage
         )
 
-        namespace = self.parser.parse_args(["-k", "keyCertSign,keyAgreement"])
+    def test_deprecated_coma_list(self) -> None:
+        """Test deprecated critical flag."""
+        msg = r"^Passing a coma-separated list is deprecated, pass space-separated values instead\.$"
+        with self.assertRemovedIn126Warning(msg):
+            namespace = self.parser.parse_args(["-k", "keyCertSign,keyAgreement"])
         self.assertEqual(
             self.key_usage(key_cert_sign=True, key_agreement=True, critical=True).value, namespace.key_usage
         )
 
     def test_deprecated_critical(self) -> None:
         """Test deprecated critical flag."""
-        msg = r"^Using critical as first value is deprecated\. The extension is critical by default\.$"
+        msg = r"^Passing a coma-separated list is deprecated, pass space-separated values instead\.$"
         with self.assertRemovedIn126Warning(msg):
             namespace = self.parser.parse_args(["-k", "critical,keyCertSign"])
-            self.assertEqual(self.key_usage(key_cert_sign=True, critical=True).value, namespace.key_usage)
+        self.assertEqual(self.key_usage(key_cert_sign=True, critical=True).value, namespace.key_usage)
 
     def test_invalid_values(self) -> None:
         """Test passing invalid values."""
         self.assertParserError(
             ["-k", "foo"],
-            "usage: {script} [-h] [-k KEY_USAGE]\n{script}: error: foo: Invalid key usage.\n",
+            "usage: {script} [-h] [-k KEY_USAGE [KEY_USAGE ...]]\n{script}: error: foo: Invalid key usage.\n",
         )
 
     def test_error(self) -> None:
         """Test wrong option values."""
         self.assertParserError(
             ["-k", "encipherOnly"],
-            "usage: {script} [-h] [-k KEY_USAGE]\n"
+            "usage: {script} [-h] [-k KEY_USAGE [KEY_USAGE ...]]\n"
             "{script}: error: encipher_only and decipher_only can only be true when key_agreement is true\n",
         )
         self.assertParserError(
             ["-k", "decipherOnly"],
-            "usage: {script} [-h] [-k KEY_USAGE]\n"
+            "usage: {script} [-h] [-k KEY_USAGE [KEY_USAGE ...]]\n"
             "{script}: error: encipher_only and decipher_only can only be true when key_agreement is true\n",
         )
 
@@ -271,47 +297,41 @@ class TLSFeatureActionTestCase(ParserTestCaseMixin, TestCase):
 
     def test_basic(self) -> None:
         """Test basic functionality of action."""
-        namespace = self.parser.parse_args(["-f", "OCSPMustStaple"])
-        self.assertEqual(
-            self.tls_feature(x509.TLSFeatureType.status_request, critical=False), namespace.tls_feature
-        )
+        namespace = self.parser.parse_args(["-f", "status_request"])
+        self.assertEqual(x509.TLSFeature([x509.TLSFeatureType.status_request]), namespace.tls_feature)
 
-        namespace = self.parser.parse_args(["-f", "critical,OCSPMustStaple"])
-        self.assertEqual(
-            self.tls_feature(x509.TLSFeatureType.status_request, critical=True), namespace.tls_feature
-        )
+        namespace = self.parser.parse_args(["-f", "status_request_v2"])
+        self.assertEqual(x509.TLSFeature([x509.TLSFeatureType.status_request_v2]), namespace.tls_feature)
 
-        namespace = self.parser.parse_args(["-f", "MultipleCertStatusRequest"])
+        namespace = self.parser.parse_args(["-f", "status_request", "status_request_v2"])
         self.assertEqual(
-            self.tls_feature(x509.TLSFeatureType.status_request_v2, critical=False), namespace.tls_feature
-        )
-
-        namespace = self.parser.parse_args(["-f", "critical,MultipleCertStatusRequest"])
-        self.assertEqual(
-            self.tls_feature(x509.TLSFeatureType.status_request_v2, critical=True), namespace.tls_feature
-        )
-
-        namespace = self.parser.parse_args(["-f", "OCSPMustStaple,MultipleCertStatusRequest"])
-        self.assertEqual(
-            self.tls_feature(
-                x509.TLSFeatureType.status_request, x509.TLSFeatureType.status_request_v2, critical=False
-            ),
+            x509.TLSFeature([x509.TLSFeatureType.status_request, x509.TLSFeatureType.status_request_v2]),
             namespace.tls_feature,
         )
 
-        namespace = self.parser.parse_args(["-f", "critical,OCSPMustStaple,MultipleCertStatusRequest"])
+    def test_deprecated_coma_list(self) -> None:
+        """Test deprecated critical flag."""
+        msg = r"^Passing a coma-separated list is deprecated, pass space-separated values instead\.$"
+        with self.assertRemovedIn126Warning(msg):
+            namespace = self.parser.parse_args(["-f", "status_request,status_request_v2"])
         self.assertEqual(
-            self.tls_feature(
-                x509.TLSFeatureType.status_request, x509.TLSFeatureType.status_request_v2, critical=True
-            ),
+            x509.TLSFeature([x509.TLSFeatureType.status_request, x509.TLSFeatureType.status_request_v2]),
             namespace.tls_feature,
         )
+
+    def test_deprecated_critical(self) -> None:
+        """Test deprecated critical flag."""
+        msg = r"^Passing a coma-separated list is deprecated, pass space-separated values instead\.$"
+        with self.assertRemovedIn126Warning(msg):
+            namespace = self.parser.parse_args(["-f", "critical,status_request"])
+        self.assertEqual(x509.TLSFeature([x509.TLSFeatureType.status_request]), namespace.tls_feature)
 
     def test_error(self) -> None:
         """Test wrong option values."""
         self.assertParserError(
             ["-f", "FOO"],
-            "usage: {script} [-h] [-f TLS_FEATURE]\n{script}: error: Unknown TLSFeature: FOO\n",
+            "usage: {script} [-h] [-f TLS_FEATURE [TLS_FEATURE ...]]\n"
+            "{script}: error: Unknown TLSFeature: FOO\n",
         )